pyproject.toml
[project]
name = "op-spectra-nn"
version = "0.1.0"
description = "Neural approximation of Eigen Functions of Differential Operators"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "comet-ml==3.54.1",
    "hydra-core==1.3.2",
    "matplotlib==3.10.7",
    "numpy==2.3.4",
    "pandas==2.3.3",
    "tensorboard==2.20.0",
    "torch==2.9.1",
    "tqdm==4.67.1",
    "transformers==4.57.1",
    "wandb==0.23.0",
]

[dependency-groups]
dev = [
    "pre-commit==4.3.0",
    "ruff==0.14.3",
    "ty==0.0.1a25",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["src"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff]
line-length = 100
